Refactor: based: activateCibXml() CRM_CHECKs that new_cib != NULL
There are two callers. The argument in startCib() is clearly non-NULL.

The argument in cib_process_command() is unsurprisingly less clear.
result_cib gets set by cib_perform_op(), and we call activateCibXml()
only if we're performing an operation that modifies the CIB, rc is
pcmk_rc_ok, and (result_cib != the_cib). result_cib should always be
non-NULL in that case. And if it ever is NULL, then this is something
I'd like to find out about as this code gets used by upstream users.

Signed-off-by: Reid Wahl <nrwahl@protonmail.com>