Refactor: based: Standardize functions in based_ipc.c
* Add Doxygen.
* Standardize log messages with what's happening in the same functions
  in other daemons.
* Standardize variable names (cib_client becomes client, op_request
  becomes msg).
* Un-indent "msg != NULL" block after moving the "msg == NULL" block
  upward and standardizing it with other daemons.
* Make sure return values and comments match what libqb documents.

The "Dispatching (un)privileged request" message is out of alignment
with the common code in other daemons. I've left it for now, because the
alternative is to fetch the client in based_ipc_dispatch_rw() and
based_ipc_dispatch_ro(), log the message, and then fetch the client
again in dispatch_common() (or pass it as an argument, which is also
non-standard).

Also add a TODO for based_ipc_destroy(). The other daemons have comments
in their DAEMON_ipc_closed() doxygen that say:

"We handle a destroyed connection the same as a closed one, but we need
a separate handler because the return type is different."

Signed-off-by: Reid Wahl <nrwahl@protonmail.com>