Feature: tools: Drop crm_attribute --inhibit-policy-engine/-i option
It have been hidden and thus informally deprecated since 051c07cc in
2009. It was never formally deprecated. However, it also looks as if it
was never documented (except that its existence was documented with no
explanation prior to 051c07cc).

It was added by commit 861fcc91 in 2005, and I haven't found anything
that used it since then (although it's impossible to be certain).

Since it's been hidden for over 16 years and I haven't found anything
that has used it, I think it's safe to drop.

Signed-off-by: Reid Wahl <nrwahl@protonmail.com>