Refactor: based: Drop config_hash
Nothing uses the values in it anymore. Prior to 1d9fb0f, we had a call
to cib_config_lookup(). (That function was removed by d59cf40.)
cib_config_lookup() looked for a value in config_hash. However, now
nothing queries it.

Previously, we were calling cib_read_config() every time the crm_config
section changed. That updated the values in config_hash and validated
them. Since nothing used it, the updates were pointless. The validation
also isn't very useful, since the options will be validated (and
defaults will be handled, etc.) when the scheduler unpacks them.

So we can take this opportunity for a little bit of simplification.

Signed-off-by: Reid Wahl <nrwahl@protonmail.com>