Refactor: libcib: Simplify read_config()
options and current_cib are now guaranteed to be non-NULL, so we don't
need to check them in a static function with one call. Also return early
if config is NULL.

g_hash_table_remove_all() is unnecessary because options is always a
new, empty table when we call this function.

pcmk__validate_cluster_options() is unnecessary because we only need the
value of one option, and pcmk__cluster_option() (called by
cib_acl_enabled()) does the same thing pcmk__validate_cluster_options()
would have done.

Signed-off-by: Reid Wahl <nrwahl@protonmail.com>