Fix: libcib: Fix setting origin attributes in CIB updates
This fixes a regression introduced by 53d4d347. The update-client
attribute was being set to the user instead of the update client.

This affects the outputs of the following public API functions and CLI
tools (via the cluster-status message with pcmk_section_times set):
* pcmk_simulate()
* pcmk_status()
* crm_mon
* crm_simulate

Also fix an indentation issue found while investigating what this
affects.

Signed-off-by: Reid Wahl <nrwahl@protonmail.com>