Refactor: libcrmcommon: pcmk__xpath_foreach_result in pcmk__apply_acls
We drop a trace log in order to do this. To get the number of matches
from pcmk__xpath_foreach_result(), we'd have to pass a struct as user
data to hold both the ACL itself and the match count. That doesn't seem
worth the trouble. We still have the trace message after applying an ACL
to a particular match.

Signed-off-by: Reid Wahl <nrwahl@protonmail.com>