Refactor: libcrmcommon: Assert arg not NULL in pcmk__unpack_acl()
And rename to pcmk__unpack_acls(). In general, we assume that an XML
node has a non-NULL doc and that its doc has a non-NULL _private field.
If we ever want to start checking that (in case a node somehow
originated outside of pcmk__xe_create()), we should do it more broadly.

The two callers already ensure that target is not NULL.

Signed-off-by: Reid Wahl <nrwahl@protonmail.com>