Refactor: libcrmcommon: Functionize resolving an ACL role reference
This involves removing some const qualifiers, because
pcmk__xpath_search() takes a non-const xmlDoc.

Our convention is to use a non-const pointer argument whenever we use
one of the argument's fields (in this case, the doc field) as a non-
const.

At first, this feels like a use case for pcmk__xe_resolve_idref(), but
that function would need to be extended in at least two ways:
* It would need to match a PCMK_XA_ID attribute instead of a
  PCMK_XA_ID_REF attribute.
* It would need to match a referenced element whose type differs from
  that of the referencing element (PCMK_XE_ACL_ROLE vs. PCMK_XE_ROLE).

Signed-off-by: Reid Wahl <nrwahl@protonmail.com>