Fix: libcrmcommon: Make parse_acl_entry() non-recursive
And rename to unpack_acl_target(). We could simplify further, except
that we've been allowing <acl_target> and <acl_group> elements to
contain <acl_permission> elements and unpacking them in violation of
the schema.

I've marked this as "fix" to add to the change log, because it does
change behavior: if <role> elements are nested within <acl_role>
elements, they will now be ignored. They should never have been
configured in the first place.

Thus far, I have made an effort to maintain backward compatibility even
for configurations that the schema doesn't allow. I have not done so
here. A workaround would risk an infinite loop of following references,
and trying to prevent that is not worth the trouble.

Signed-off-by: Reid Wahl <nrwahl@protonmail.com>