Refactor: libcrmcommon: Use foreach for adding attr changes to patchset
This change is larger than I'd prefer, but everything here is tied
together and I think review will be easier if we do it in one step.

In particular, I thought that the previous "create the change element
if it's NULL" on every iteration if we found a changed attribute, made
the whole thing harder to read. So at the cost of a bit of extra memory
allocation (for a GSList) and iteration, we move to the following
approach:

1. Make a list of changed or deleted attributes. If the resulting list
   is empty, stop. Otherwise, continue to next steps.
2. Create a PCMK_XE_CHANGE child of the patchset.
3. Create a PCMK_XE_CHANGE_LIST child of the PCMK_XE_CHANGE.
4. For each changed or deleted attribute, add a PCMK_XE_CHANGE_ATTR
   child to the PCMK_XE_CHANGE_LIST.
5. Create a PCMK_XE_CHANGE_RESULT child of the PCMK_XE_CHANGE.
6. Create an xml->name child of the PCMK_XE_CHANGE_RESULT.
7. For each attribute that was not deleted, copy it from the source xml
   to the xml->name child.
8. Free the list of changed attributes.

Signed-off-by: Reid Wahl <nrwahl@protonmail.com>