Low: libcrmcommon: Drop "<null>" fallback in show_xml_element()
If show_xml_element() wanted to show an attribute with a NULL value, it
would use "<null>" for the value. This didn't make sense. Immediately
prior, we XML-escaped the value. "<null>" has XML special characters and
would need to be escaped in order to be valid. (Of course,
pcmk__xml_show() and its helpers are currently used only with text-like
output formats; for those, having perfectly valid XML is less
important.)

We could use a different fallback string, such as "(null)". However, for
any XML element that we create, every attribute in its properties list
should have a non-NULL value. So this should have little or no practical
effect. "<null>" was a fallback in case of bugs, strange inputs via the
public API (for example, do_crm_log_xml()), or other unexpected
scenarios. As far as I know, we never promised that NULL attributes
would be represented in the output. So I prefer to simply not show them.

This will soon allow us to call pcmk__xml_dump_attr() in
show_xml_element().

Signed-off-by: Reid Wahl <nrwahl@protonmail.com>