Refactor: libcrmcommon: Drop a couple checks from show_xml_element()
An attribute's value really shouldn't be NULL unless the XML came from
public API (and even then, it would be strange). pcmk__dump_xml_attr()
already ignores attributes with NULL values. So dropping the NULL check
simply means that if an attribute's name is in PCMK__XA_HIDDEN and its
value is NULL, we'll now show "*****" for its value instead of skipping
it. That should basically never happen in practice, and if it does, this
behavior is arguably more correct anyway.

Also, an attribute's name can't be empty in valid XML.

Signed-off-by: Reid Wahl <nrwahl@protonmail.com>