Refactor: libcrmcommon: Functionize ACL-filtering one match
Note to reviewer: this code is expected to be removed a few commits from
now. This commit was part of my process of trying to understand the
filtering logic. It may make review easier, or it may make it more
tedious.

Use pcmk__xpath_foreach_result(). Also drop a trace message, for two
reasons:
* We don't have easy access to num_results anymore. (We'd have to get it
  via a counter variable or something.)
* The message tells us how many nodes were filtered by a given XPath,
  but not which nodes. So it doesn't seem especially useful. If we want
  tracing for ACL filtering (which seems like a reasonable thing to
  want), then we should make it more granular so that it's informative.

Signed-off-by: Reid Wahl <nrwahl@protonmail.com>