Refactor: libcrmcommon: Clarify xml_acl_filtered_copy()
This is a substantial refactor. Most of the details are in the comments.

It took me days to understand exactly what xml_acl_filtered_copy() was
doing, particularly in the calls to purge_xml_attributes(). For a while
I was pretty sure that we could be exposing "denied" children if an
ancestor had the pcmk__xf_acl_read or pcmk__xf_acl_write flag set. I'm
no longer confident that was the case, although it may have been.

The whole process was convoluted. We unpacked all of a user's ACLs and
applied them to matching nodes throughout the document. Then we went
through the unpacked list of a user's ACLs (docpriv->acls) and found the
ones with "deny" permissions and non-NULL xpath. (By the way, I believe
all ACLs should have non-NULL xpath by that point.)

Then for each of those "deny" ACLs, we filtered the document based on
it. We ran an XPath query to find all the nodes in the document that
matched the "deny" ACL, and we called purge_xml_attributes() on each
such node. A return value of true indicated that the current node either
was readable or had at least one readable descendant.

* If the current node had the pcmk__xf_acl_read or pcmk__xf_acl_write
  flag set (is_mode_allowed()), then it was readable and we returned
  true immediately. Some read or write ACL had matched that node during
  the apply stage. Of course this could never be true for the top-level
  call, because the top-level call was always for some node that a
  "deny" ACL had matched.

* Otherwise, the current node was unreadable. We called
  purge_xml_attributes() on all children.

  - If any children returned true, then we had at least one readable
    descendant. We removed all attributes except for ID and then
    returned true.

  - Otherwise, we had no readable descendants. Since the current node
    was also unreadable, we removed the current node's subtree and
    returned false.

As mentioned, this was convoluted. The key observation is that before we
did any of that, we had already applied all the ACLs throughout the
document. So all we have to do is recurse from the top down, determine
which nodes are readable for the current ACL user, and filter out the
rest. The comments explain how this works.

Signed-off-by: Reid Wahl <nrwahl@protonmail.com>