Refactor: libcib: New process_create_xpath()
I don't mind a bit of duplication for now, to avoid a monolithic
function. This is mostly just splitting out the op-specific logic. We'll
make further improvements soon.

We do go ahead and take advantage of the fact that process_xpath()
always broke out of the for-loop after the first iteration for create
ops.

Signed-off-by: Reid Wahl <nrwahl@protonmail.com>