Refactor: libcrmcommon: pcmk__output_xml_create_parent() drops list arg
I don't think the variadic argument helps readability much. It avoids
the need to store the result in an xmlNode * variable in the caller,
which can save a line. But it requires an extra line for the NULL
terminator, and it requires indenting each line much farther, which is
one reason we have so many temp variables.

Signed-off-by: Reid Wahl <nrwahl@protonmail.com>