Refactor: libcib: Allow cib__process_apply_patch() to use existing_cib
Previously, we made a copy no matter what.

There are only two callers: cib_perform_op() and
cib_apply_patch_event().

When called by cib_perform_op(), it is always the case that either
*result_cib == existing_cib or *result_cib is a copy of existing_cib. In
either case, we don't need to make another copy.

cib_apply_patch_event(), on the other hand, can pass an arbitrary value
for result_cib, since it's a public API function. In order to keep sane
behavior, we free the result CIB and set it to a copy of the existing
CIB if it doesn't point to the existing CIB initially.

Signed-off-by: Reid Wahl <nrwahl@protonmail.com>