Refactor: libcib: Free *result_cib unconditionally in cib__process_erase()
*result_cib is always either existing_cib itself or a copy of
existing_cib. When we run these operations on existing_cib (without
making a copy), the whole point is that we want existing_cib to reflect
the changes when we're finished.

So in this case, we always want *result_cib to be replaced by an empty
CIB, starting with the original version numbers and then bumping admin
epoch.

Signed-off-by: Reid Wahl <nrwahl@protonmail.com>