Refactor: libcib: Merge existing_cib and result_cib args of cib__op_fn_t
Immediately prior to this commit:
- *result_cib was always either the existing CIB or a copy of it.
- cib__process_query() was the only cib__op_fn_t that still used its
  existing_cib argument. It did not use its result_cib argument.

If *result_cib was a copy of the existing CIB, then we didn't want
changes to *result_cib to affect the existing CIB until the operation
succeeded. If *result_cib was the existing CIB itself, then we didn't
have a copy, and we wanted the changes to apply directly to the existing
CIB.

We maintain that behavior in this commit.

Signed-off-by: Reid Wahl <nrwahl@protonmail.com>