Refactor: libcib: Add digest in cib_perform_op if and only if tracing
Checking the digest once per minute is an apparently arbitrary interval
for sanity-checking. However, we don't actually USE the digest and do
the check unless we're tracing (although we would log it at info level
prior to this commit). Since we're almost never tracing, only add the
digest if we're tracing. This lets us drop the expires/tm_now code.

This functionality was added by commit 6a2ffe20 with no explanation
beyond what's in the code comment.

Signed-off-by: Reid Wahl <nrwahl@protonmail.com>