Refactor: libcib: Drop test apply-patchset when tracing cib_perform_op()
In theory, this could be a nice sanity-check, to catch a mistake if we
mess up xml_apply_patchset() or something related.

In practice, we never enable tracing for cib_perform_op(), so we would
never actually SEE an error here. By the time we've enabled tracing, we
already know there's a problem. At that point, we can add tracing and
recompile, or run under gdb, or whatever other debugging strategy we
prefer. We don't need to leave all this tracing lying around just in
case, when we're unlikely to ever use it.

This tracing was introduced over the course of a few commits. c96c99ae
and c55115d5 look like the main relevant ones. This may have been more
relevant or useful when we v2 patchsets were being developed and we may
have been uncertain that they would work.

Also drop a comment about v2 patchsets that probably hasn't been
relevant since we dropped support for rolling upgrades from Pacemaker
versions that create v1 patchsets. We only support v2.

Signed-off-by: Reid Wahl <nrwahl@protonmail.com>